Updated some files
PlaceIT now really tests the npc's presence in the room. This is done because for some reason, the game returns an empty array when a monster is in the room. A test for the monsters will be added later.
Command now has ME. It's a new command used to see our current state (hp, inventory, etc.)
Modified Place's get functions to work the same as inventory
Player now as a toString function
Added Me command to Game, added TODO to check why the list is empty with monsters

Signed-off-by: Creerio <10045021-Creerio@users.noreply.gitlab.com>